move well-known integration tests to their own file
